Moves all court components into seperate components folder
